refactor: extract common property fields
